Adding libspy.
It needs to be used by smman as this will consume less cpu cycles than
managing inotify events. When having a lot of logs, inotify events gets
too annoying, thus the fact that we miss some of them.

With libspy, files will be polled periodically, no matter the activity
on them.