Fix an infinite loop if elasticsearch goes crazy.
We had a case here where elasticsearch stopped answering to HTTP queries
SMMAN was sending.

This led SMMAN to a state where he had 1015 sockets opened to
elasticsearch, and constantly trying to open logfiles as they had new
logs in it, but unable to open them because we reached the 1024 nofile
limit.

So this patch change behavior :
- Add a 10s timeout to ecore_con_url objects
- Dont instantly stat() files right after processing if an error
  happenned at last try.

This bug happenned twice here with a server sending over 1k logs in the
same second.